requirements.txt
# FastAPI and Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Data Processing
numpy==2.1.1
pandas==2.2.2
polars==1.9.0

# Machine Learning
scikit-learn==1.5.2
lightgbm==4.5.0
optuna==3.6.1
joblib==1.4.2

# Database
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.2 # ADDED

# Redis
redis==5.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# CLI
typer==0.12.5

# Environment
python-dotenv==1.0.1

# Logging
python-json-logger==2.0.7